streamline autograd function wrapping
