+++ thunder/core/jit_ext.py
@@ -921,13 +921,9 @@ def _general_jit_torch_autograd_function_apply_lookaside(obj: Any, *args, **kwar
 
     custom_autograd_function_cls = unwrap(obj)
     custom_forward = custom_autograd_function_cls.forward
-    args_, kwargs_ = tree_map(unwrap, (args, kwargs))
     ctx = torch.autograd.function.FunctionCtx()
-
-    pr = ProvenanceRecord(PseudoInst.LOOKASIDE, inputs=[wrap_const(custom_forward).provenance])
-    wrapped_ctx = wrap(ctx, provenance=pr)
-    args_, kwargs_ = tree_map(lambda a: wrap(a, provenance=pr), (args_, kwargs_))
-    return _interpret_call(custom_forward, wrapped_ctx, *args_, **kwargs_)
+    wrapped_ctx = wrap_const(ctx)
+    return _interpret_call(custom_forward, wrapped_ctx, *args, **kwargs)
 
 
 @register_general_jit_lookaside(torch.finfo)
